requirements.txt
fastapi 
uvicorn 
httpx 
python-dotenv
cryptography
passlib[bcrypt]
python-jose[cryptography]